+++ src/hash_index.rs
@@ -315,7 +315,7 @@ where
     {
         let barrier = Barrier::new();
         let hash = self.hash(key);
-        if let Ok(Some((mut locker, data_block_mut, mut entry_ptr))) =
+        if let Ok(Some((mut locker, data_block_mut, mut entry_ptr, _))) =
             self.get_entry(key, hash, &mut (), &barrier)
         {
             let (k, v) = entry_ptr.get(data_block_mut);
@@ -389,7 +389,7 @@ where
             {
                 let barrier = Barrier::new();
                 if let Ok(result) = self.get_entry(key, hash, &mut async_wait_pinned, &barrier) {
-                    if let Some((mut locker, data_block_mut, mut entry_ptr)) = result {
+                    if let Some((mut locker, data_block_mut, mut entry_ptr, _)) = result {
                         let (k, v) = entry_ptr.get(data_block_mut);
                         let modify_action: ModifyAction<V> = updater(k, v).into();
                         let result = match modify_action {
@@ -451,7 +451,7 @@ where
         F: FnOnce(&K, &mut V) -> R,
     {
         let barrier = Barrier::new();
-        let (mut locker, data_block_mut, mut entry_ptr) = self
+        let (mut locker, data_block_mut, mut entry_ptr, _) = self
             .get_entry(key, self.hash(key), &mut (), &barrier)
             .ok()
             .flatten()?;
@@ -492,7 +492,7 @@ where
             let mut async_wait = AsyncWait::default();
             let mut async_wait_pinned = Pin::new(&mut async_wait);
             if let Ok(result) = self.get_entry(key, hash, &mut async_wait_pinned, &Barrier::new()) {
-                if let Some((mut locker, data_block_mut, mut entry_ptr)) = result {
+                if let Some((mut locker, data_block_mut, mut entry_ptr, _)) = result {
                     let (k, v) = entry_ptr.get_mut(data_block_mut, &mut locker);
                     return Some(updater(k, v));
                 }