+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-1.6.4
+1.7.0
 
-* Optimize `Hash*::update*`.
-* Work-in-progress `HashCache` implementation.
+* Optimize `Hash*::update*` and `HashIndex::modify*`.
+* API update 1: _WORK IN PROGRESS_ `HashCache` implementation.
+* API update 2: add `HashMap::{get, get_async}` returning an `OccupiedEntry`.
 
 1.6.3